backend/requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.14
fonttools==4.58.5
geographiclib==2.0
geopandas==1.1.1
geopy==2.4.1
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mercantile==1.2.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pystac==1.13.0
pystac-client==0.8.6
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
rpds-py==0.26.0
scipy==1.16.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
whitebox==2.3.6
xyzservices==2025.4.0
python-dotenv